requirements.txt
LM_Cocktail==0.0.4
faiss-gpu==1.7.2
numpy==1.26.4
FlagEmbedding==1.2.9
torch==2.2.1+cu118
transformers==4.40.2
tqdm==4.66.4